sore clean
Former-commit-id: cb13818cef7bf020201f1c3b906f1318dbeed13c

+++ lib/ncurses/src/Graphical/Window.cpp
@@ -113,7 +113,7 @@ int Window::getColorPair(Color fg, Color bg)
 
 void Window::drawBufferPixel(int x, int y, Color color)
 {
-    if (y >= _height || x >= _width || x < 0 || y < 0 || color.a == 0)
+    if ((size_t)y >= _height || (size_t)x >= _width || (size_t)x < 0 || (size_t)y < 0 || color.a == 0)
         return;
     _colorBuffer[y][x] = color;
 }
